+++ lume/LumeScene/src/asset/asset_object.h
@@ -30,7 +30,7 @@ class IAssetObject : public CORE_NS::IInterface {
     META_INTERFACE(CORE_NS::IInterface, IAssetObject, "70c77b11-bd85-490a-880f-d8a80d2addb0")
 public:
     virtual bool Load(const IScene::Ptr&, BASE_NS::string_view uri, bool createResources,
-        const CORE_NS::ResourceId& rid, size_t offset) = 0;
+        const CORE_NS::ResourceId& rid, int64_t offset) = 0;
 };
 
 META_REGISTER_CLASS(AssetObject, "a8b4c9e9-9c28-49b7-85b2-2eb6ac812b7c", META_NS::ObjectCategoryBits::NO_CATEGORY)
@@ -40,7 +40,7 @@ class AssetObject : public META_NS::IntroduceInterfaces<META_NS::MetaObject, IAs
 
 public:
     bool Load(const IScene::Ptr&, BASE_NS::string_view uri, bool createResources, const CORE_NS::ResourceId& rid,
-        size_t offset) override;
+        int64_t offset) override;
 
 private:
     void CreateImageResources(
